requirements.txt
django>=1.6,<1.7
djangorestframework
django-extensions
ipython
